Document default parameters for escapeMapping
This allows users to copy and extend the existing parameters. If the users have a working document with the default parameters but have to add a rule because of a new document with different texts they won't know what to write as default rules (that might be still needed for the first document) otherwise.